+++ packages/route/_test_deps.ts
@@ -1,1 +1,0 @@
-export * from "@std/assert";